refactor: split tau-tools tools.rs into policy/session/bash modules (#1510)
